+++ database/migrations/2025_04_27_164647_create_historynotas_table.php
@@ -11,19 +11,21 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('historynotas', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('transaction_id');
-            $table->unsignedBigInteger('customer_id');
-            $table->string('nota_file');
-            $table->timestamp('tanggal_cetak')->nullable();
-            $table->tinyInteger('deleteSts')->default(0);
-            $table->string('createdBy')->nullable();
-            $table->string('updatedBy')->nullable();
-            $table->timestamps();
-            $table->foreign('transaction_id')->references('id')->on('transactions')->onDelete('cascade');
-            $table->foreign('customer_id')->references('id')->on('customers')->onDelete('cascade');
-        });
+        if (!Schema::hasTable('historynotas')) {
+            Schema::create('historynotas', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('transaction_id');
+                $table->unsignedBigInteger('customer_id');
+                $table->string('nota_file');
+                $table->timestamp('tanggal_cetak')->nullable();
+                $table->tinyInteger('deleteSts')->default(0);
+                $table->string('createdBy')->nullable();
+                $table->string('updatedBy')->nullable();
+                $table->timestamps();
+                $table->foreign('transaction_id')->references('id')->on('transactions')->onDelete('cascade');
+                $table->foreign('customer_id')->references('id')->on('customers')->onDelete('cascade');
+            });
+        }
     }
 
     /**
